+++ ndn/experiments/experiment.py
@@ -35,6 +35,8 @@ class Experiment:
         self.nPings = args["nPings"]
         self.strategy = args["strategy"]
         self.pctTraffic = args["pctTraffic"]
+        self.nlsrSecurity = args["nlsrSecurity"]
+        self.arbArgs = args["arbArgs"]
 
         # Used to restart pings on the recovered node if any
         self.pingedDict = {}