+++ ndn/ndn_host.py
@@ -102,9 +102,7 @@ class NdnHost(Host, NdnHostCommon):
         self.cmd("mkdir -p %s" % self.homeFolder)
         self.cmd("cd %s" % self.homeFolder)
 
-        self.nfd = Nfd(self)
-        print "ndn_host.py--------------------------Ndnhost() cla---- init"
-        self.nfd.start()
+        self.nfd = None
 
         self.peerList = {}
 
@@ -119,7 +117,8 @@ class NdnHost(Host, NdnHostCommon):
 
     def terminate(self):
         "Stop node."
-        self.nfd.stop()
+        if self.nfd is not None:
+            self.nfd.stop()
         Host.terminate(self)
 
 class CpuLimitedNdnHost(CPULimitedHost, NdnHostCommon):
@@ -137,8 +136,7 @@ class CpuLimitedNdnHost(CPULimitedHost, NdnHostCommon):
         self.cmd("mkdir -p %s" % self.homeFolder)
         self.cmd("cd %s" % self.homeFolder)
 
-        self.nfd = Nfd(self)
-        self.nfd.start()
+        self.nfd = None
 
         self.peerList = {}
 
@@ -153,8 +151,9 @@ class CpuLimitedNdnHost(CPULimitedHost, NdnHostCommon):
 
     def terminate(self):
         "Stop node."
-        self.nfd.stop()
-        Host.terminate(self)
+        if self.nfd is not None:
+            self.nfd.stop()
+        CPULimitedHost.terminate(self)
 
 class NdnStation(Station, NdnHostCommon):
     "NDNStation is a Host that always runs NFD, and is wifi-enabled"